app/requirements.txt
moviepy==2.1.2
PyYAML==6.0.2
pillow==10.4.0
